Move blur functions to separate module
